copy static files if not exist

+++ app/builder.js
@@ -47,7 +47,10 @@ confArray.forEach(confObj => {
         // copies static files in output
         let staticFiles = resource.getStaticFiles(dt);
         staticFiles.forEach(f => {
-            fs.writeFileSync(outputPath + '/' + f.filename, f.file);
+            if (fs.existsSync(outputPath + '/' + f.filename) === false) {
+                fs.writeFileSync(outputPath + '/' + f.filename, f.file);
+                console.log(`${log_filename_tag} file: ${f.filename} COPIED!`);
+            }
         });
 
         fs.writeFileSync(outputPath + '/' + jsonObj.config.css, resource.getCss(dt));
